Update rgbww_light_output.h
Changed cold_white_temperature and warm_white temperature from a float to an integer. This makes home assisting show it correctly in the color temperature slider. If it is set to float home assistant show something like 470.8390000283847829304845 in the slider which provides an ugly ui and color temperature to as a decimal is invalid anyway.

+++ esphome/components/rgbww/rgbww_light_output.h
@@ -46,8 +46,8 @@ class RGBWWLightOutput : public light::LightOutput {
   output::FloatOutput *blue_;
   output::FloatOutput *cold_white_;
   output::FloatOutput *warm_white_;
-  float cold_white_temperature_;
-  float warm_white_temperature_;
+  int cold_white_temperature_;
+  int warm_white_temperature_;
   bool constant_brightness_;
   bool color_interlock_{false};
 };
